[ref] renamed aux -> aux_ to support windows
AUX is a reserved name in windows.
