+++ src/makefile
@@ -8,8 +8,8 @@ SOEXT=so
 
 SODIUM_NEWER_THAN_1_0_18 := $(shell pkgconf --atleast-version=1.0.19 libsodium; echo $$?)
 ifeq ($(SODIUM_NEWER_THAN_1_0_18),1)
-	CFLAGS+= -Iaux
-	EXTRA_OBJECTS+= aux/kdf_hkdf_sha512.o
+	CFLAGS+= -Iaux_
+	EXTRA_OBJECTS+= aux_/kdf_hkdf_sha512.o
 else
 	CFLAGS+= -DHAVE_SODIUM_HKDF=1
 endif
@@ -69,7 +69,7 @@ $(PREFIX)/bin/opaque: utils/opaque
 clean:
 	rm -f \
 		*.o \
-		aux/*.o \
+		aux_/*.o \
 		libopaque.dll \
 		libopaque.so \
 		tests/opaque-munit \
